Simplify and generalize convert df to numpy for time series and train them all memory
