+++ tests/data/climate_small_schema.json
@@ -1,0 +1,35 @@
+{
+  "measure_date": "<M8[ns]",
+  "stations_id": "int",
+  "fx": "float",
+  "fm": "float",
+  "rsk": "float",
+  "rskf": "float",
+  "sdk": "float",
+  "shk_tag": "float",
+  "nm": "float",
+  "vpm": "float",
+  "pm": "float",
+  "tmk": "float",
+  "upm": "float",
+  "txk": "float",
+  "tnk": "float",
+  "tgk": "float",
+  "rs": "float",
+  "rsf": "float",
+  "sh_tag": "float",
+  "nsh_tag": "float",
+  "v_te002m": "float",
+  "v_te005m": "float",
+  "v_te010m": "float",
+  "v_te050m": "float",
+  "atmo_strahl": "float",
+  "fd_strahl": "float",
+  "fg_strahl": "float",
+  "ash_6": "float",
+  "wash_6": "float",
+  "waas_6": "float",
+  "stations_height": "float",
+  "geo_width": "float",
+  "geo_length": "float"
+}
